C# SDK (#1)
* committing stashed changes from temp work while away.

* Updating handlers and topic_manager to take context for sqlite data transactions.

* updating tests mockTopicManager in handlers_test.go to conform to updated TopicManager interface that takes context.Context from calling handler.

* Updated sqlite storage to have a newSqliteStorage method for getting an implementation of Storage interface for SQLite, and added it to the NewStorage factory method to get a storage based on the configuration.

* removing the main binary from version control

* Adding .NET SDK.
Adding base layout for the C# usage library, specifically the interface for messaging and builder that will create the messaging client.
Removed the CLI part of the project since it is not needed (at least for now) and wanted to clean things up.

* Implemented C# SDK.
Created handlers for every type of message to the server in the C# sdk, with custom exceptions.

* Removed the obj files from the repo.

* Modifying .gitignore to include proper obj and bin ignoring.